Move multisig key encoder into database
This encoder is primarily used in the DB, and will be deleted from being
used in Bech32 Encoder soon.
