+++ ironfish/src/wallet/account/account.ts
@@ -14,7 +14,6 @@ import { StorageUtils } from '../../storage/database/utils'
 import { WithNonNull, WithRequired } from '../../utils'
 import { DecryptedNote } from '../../workerPool/tasks/decryptNotes'
 import { AssetBalances } from '../assetBalances'
-import { isSignerMultisig } from '../exporter/multisigKeys'
 import { MultisigKeys, MultisigSigner } from '../interfaces/multisigKeys'
 import { WalletBlockHeader } from '../remoteChainProcessor'
 import { AccountValue } from '../walletdb/accountValue'
@@ -22,6 +21,7 @@ import { AssetValue } from '../walletdb/assetValue'
 import { BalanceValue } from '../walletdb/balanceValue'
 import { DecryptedNoteValue } from '../walletdb/decryptedNoteValue'
 import { HeadValue } from '../walletdb/headValue'
+import { isSignerMultisig } from '../walletdb/multisigKeys'
 import { TransactionValue } from '../walletdb/transactionValue'
 import { WalletDB } from '../walletdb/walletdb'
