+++ ironfish/src/wallet/exporter/bech32.ts
@@ -6,8 +6,8 @@ import bufio, { EncodingError } from 'bufio'
 import { Bech32m } from '../../utils'
 import { ACCOUNT_SCHEMA_VERSION } from '../account/account'
 import { AccountImport, KEY_LENGTH, VIEW_KEY_LENGTH } from '../walletdb/accountValue'
+import { MultisigKeysEncoding } from '../walletdb/multisigKeys'
 import { AccountDecodingOptions, AccountEncoder, DecodeFailed, DecodeInvalid } from './encoder'
-import { MultisigKeysEncoding } from './multisigKeys'
 
 export const BECH32_ACCOUNT_PREFIX = 'ifaccount'
